fix: move recommend wallet to scence
